pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "b1zn3",email = "141233108+B1zn3@users.noreply.github.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.1,<0.130.0",
    "uvicorn>=0.41.0,<0.42.0",
    "sqlalchemy>=2.0.46,<3.0.0",
    "pydantic-settings>=2.13.1,<3.0.0",
    "alembic>=1.18.4,<2.0.0",
    "python-jose>=3.5.0,<4.0.0",
    "redis (>=7.2.0,<8.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
]

[build-system]
requires = ["setuptools>=61.0"]          
build-backend = "setuptools.build_meta"  